katetextlayout.h needs to include katelinerange.h otherwise gcc-3.3 says kdelibs/kdecore/util/ksharedptr.h: In destructor `KSharedPtr<T>::~KSharedPtr() [with T = KateLineLayout]': kdelibs/kate/part/katetextlayout.h:100:   instantiated from here kdelibs/kdecore/util/ksharedptr.h:77: error: invalid use of undefined type `struct KateLineLayout' But katelinerange.h doesn't need to include katetextlayout.h.
svn path=/trunk/KDE/kdelibs/kate/; revision=630954

+++ part/part/katejscript.cpp
@@ -668,7 +668,7 @@ JSValue* KateJSDocumentProtoFunc::callAsFunction(KJS::ExecState *exec,
 
     case KateJSDocument::LastIndexOf: {
       if (exception.invalidArgs(2, 4)) break;
-      const int from = args.size() >=3 ? args[2]->toUInt32(exec) : -1;
+      const int from = args.size() >=3 ? (int)args[2]->toUInt32(exec) : -1;
       Qt::CaseSensitivity cs = Qt::CaseSensitive;
       if (args.size() == 4 && !args[3]->toBoolean(exec))
         cs = Qt::CaseInsensitive;
